Updated login.html as of Feb 21
Line 211:  Changed the error message from "Please fix the errors before submitting." to "Please enter valid login credentials."
Line 179 to 191: If else statements to check the submitButtonError and emailError display while username text-field is in focus. If display is "block", change it to "none".
Line 193 to 205: If else statements to check the submitButtonError and emailError display while password text-field is in focus. If display is "block", change it to "none".